pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-generativeai>=0.8.5",
    "pandas>=2.3.0",
    "requests>=2.32.3",
    "streamlit>=1.45.1",
    "trafilatura>=2.0.0",
]
